// classifier/backend/src/types_pb.ts
export enum Confidence {
  /**
   * CSV empty + No guess. UI: User must select.
   *
   * @generated from enum value: CONFIDENCE_UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   * CSV empty + Heuristic guess. UI: User must confirm.
   *
   * @generated from enum value: CONFIDENCE_UNCERTAIN = 1;
   */
  UNCERTAIN = 1,

  /**
   * Value in CSV. UI: Pre-filled, user can skip.
   *
   * @generated from enum value: CONFIDENCE_CERTAIN = 2;
   */
  CERTAIN = 2,
}
// ...
export enum Track {
  /**
   * @generated from enum value: TRACK_UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   * @generated from enum value: TRACK_RAINBOW_ROAD = 1;
   */
  RAINBOW_ROAD = 1,
}
// Retrieve enum metadata with: proto3.getEnumType(Track)
proto3.util.setEnumType(Track, "mkw.shared.Track", [
  { no: 0, name: "TRACK_UNKNOWN" },
  { no: 1, name: "TRACK_RAINBOW_ROAD" },
]);

/**
 * @generated from message mkw.shared.RaceMetadata
 */
export class RaceMetadata extends Message<RaceMetadata> {
  /**
   * @generated from field: mkw.shared.Track track = 1;
   */
  track = Track.UNKNOWN;

  /**
   * @generated from field: int32 total_laps = 2;
   */
  totalLaps = 0;

  constructor(data?: PartialMessage<RaceMetadata>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "mkw.shared.RaceMetadata";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "track", kind: "enum", T: proto3.getEnumType(Track) },
    { no: 2, name: "total_laps", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): RaceMetadata {
    return new RaceMetadata().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): RaceMetadata {
    return new RaceMetadata().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): RaceMetadata {
    return new RaceMetadata().fromJsonString(jsonString, options);
  }

  static equals(a: RaceMetadata | PlainMessage<RaceMetadata> | undefined, b: RaceMetadata | PlainMessage<RaceMetadata> | undefined): boolean {
    return proto3.util.equals(RaceMetadata, a, b);
  }
}

/**
 * Defines a Backend Heuristic Guess
 *
 * @generated from message mkw.shared.TelemetryGuess
 */
export class TelemetryGuess extends Message<TelemetryGuess> {
  /**
   * @generated from field: mkw.shared.Confidence confidence = 1;
   */
  confidence = Confidence.UNKNOWN;

  /**
   * @generated from oneof mkw.shared.TelemetryGuess.value
   */
  value: {
    /**
     * @generated from field: mkw.shared.Placement placement = 2;
     */
    value: Placement;
    case: "placement";
  } | {
    /**
     * @generated from field: mkw.shared.LapCount lap_count = 3;
     */
    value: LapCount;
    case: "lapCount";
  } | {
    /**
     * @generated from field: int32 coin_count = 4;
     */
    value: number;
    case: "coinCount";
  } | {
    /**
     * @generated from field: mkw.shared.ItemType item = 5;
     */
    value: ItemType;
    case: "item";
  } | {
    /**
     * @generated from field: mkw.shared.RacePhase race_phase = 6;
     */
    value: RacePhase;
    case: "racePhase";
  } | {
    /**
     * @generated from field: mkw.shared.Track track = 7;
     */
    value: Track;
    case: "track";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<TelemetryGuess>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "mkw.shared.TelemetryGuess";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "confidence", kind: "enum", T: proto3.getEnumType(Confidence) },
    { no: 2, name: "placement", kind: "enum", T: proto3.getEnumType(Placement), oneof: "value" },
    { no: 3, name: "lap_count", kind: "enum", T: proto3.getEnumType(LapCount), oneof: "value" },
    { no: 4, name: "coin_count", kind: "scalar", T: 5 /* ScalarType.INT32 */, oneof: "value" },
    { no: 5, name: "item", kind: "enum", T: proto3.getEnumType(ItemType), oneof: "value" },
    { no: 6, name: "race_phase", kind: "enum", T: proto3.getEnumType(RacePhase), oneof: "value" },
    { no: 7, name: "track", kind: "enum", T: proto3.getEnumType(Track), oneof: "value" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): TelemetryGuess {
    return new TelemetryGuess().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): TelemetryGuess {
    return new TelemetryGuess().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): TelemetryGuess {
    return new TelemetryGuess().fromJsonString(jsonString, options);
  }

  static equals(a: TelemetryGuess | PlainMessage<TelemetryGuess> | undefined, b: TelemetryGuess | PlainMessage<TelemetryGuess> | undefined): boolean {
    return proto3.util.equals(TelemetryGuess, a, b);
  }
}
// ...
export class GetFrameClassificationResponse extends Message<GetFrameClassificationResponse> {
  /**
   * @generated from field: string frame_id = 1;
   */
  frameId = "";

  /**
   * @generated from field: string image_url = 2;
   */
  imageUrl = "";

  /**
   * @generated from field: mkw.shared.RaceMetadata metadata = 3;
   */
  metadata?: RaceMetadata;

  /**
   * @generated from field: mkw.shared.TelemetryGuess placement = 4;
   */
  placement?: TelemetryGuess;

  /**
   * @generated from field: mkw.shared.TelemetryGuess lap_count = 5;
   */
  lapCount?: TelemetryGuess;

  /**
   * @generated from field: mkw.shared.TelemetryGuess coin_count = 6;
   */
  coinCount?: TelemetryGuess;

  /**
   * @generated from field: mkw.shared.TelemetryGuess primary_item = 7;
   */
  primaryItem?: TelemetryGuess;

  /**
   * @generated from field: mkw.shared.TelemetryGuess secondary_item = 8;
   */
  secondaryItem?: TelemetryGuess;

  /**
   * @generated from field: mkw.shared.TelemetryGuess race_phase = 9;
   */
  racePhase?: TelemetryGuess;

  constructor(data?: PartialMessage<GetFrameClassificationResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "mkw.shared.GetFrameClassificationResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "frame_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "image_url", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "metadata", kind: "message", T: RaceMetadata },
    { no: 4, name: "placement", kind: "message", T: TelemetryGuess },
    { no: 5, name: "lap_count", kind: "message", T: TelemetryGuess },
    { no: 6, name: "coin_count", kind: "message", T: TelemetryGuess },
    { no: 7, name: "primary_item", kind: "message", T: TelemetryGuess },
    { no: 8, name: "secondary_item", kind: "message", T: TelemetryGuess },
    { no: 9, name: "race_phase", kind: "message", T: TelemetryGuess },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetFrameClassificationResponse {
    return new GetFrameClassificationResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GetFrameClassificationResponse {
    return new GetFrameClassificationResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GetFrameClassificationResponse {
    return new GetFrameClassificationResponse().fromJsonString(jsonString, options);
  }

  static equals(a: GetFrameClassificationResponse | PlainMessage<GetFrameClassificationResponse> | undefined, b: GetFrameClassificationResponse | PlainMessage<GetFrameClassificationResponse> | undefined): boolean {
    return proto3.util.equals(GetFrameClassificationResponse, a, b);
  }
}

/**
 * 2. Sending the guesses back
 *
 * @generated from message mkw.shared.SaveFrameClassificationRequest
 */
export class SaveFrameClassificationRequest extends Message<SaveFrameClassificationRequest> {
  /**
   * @generated from field: string frame_id = 1;
   */
  frameId = "";

  /**
   * @generated from field: mkw.shared.Placement placement = 2;
   */
  placement = Placement.PLACE_UNKNOWN;

  /**
   * @generated from field: mkw.shared.LapCount lap_count = 3;
   */
  lapCount = LapCount.LAP_UNKNOWN;

  /**
   * @generated from field: optional int32 coin_count = 4;
   */
  coinCount?: number;

  /**
   * @generated from field: mkw.shared.ItemType primary_item = 5;
   */
  primaryItem = ItemType.ITEM_NONE;

  /**
   * @generated from field: mkw.shared.ItemType secondary_item = 6;
   */
  secondaryItem = ItemType.ITEM_NONE;

  /**
   * @generated from field: mkw.shared.RacePhase race_phase = 7;
   */
  racePhase = RacePhase.PHASE_UNKNOWN;

  /**
   * @generated from field: mkw.shared.Track track = 8;
   */
  track = Track.UNKNOWN;

  constructor(data?: PartialMessage<SaveFrameClassificationRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "mkw.shared.SaveFrameClassificationRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "frame_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "placement", kind: "enum", T: proto3.getEnumType(Placement) },
    { no: 3, name: "lap_count", kind: "enum", T: proto3.getEnumType(LapCount) },
    { no: 4, name: "coin_count", kind: "scalar", T: 5 /* ScalarType.INT32 */, opt: true },
    { no: 5, name: "primary_item", kind: "enum", T: proto3.getEnumType(ItemType) },
    { no: 6, name: "secondary_item", kind: "enum", T: proto3.getEnumType(ItemType) },
    { no: 7, name: "race_phase", kind: "enum", T: proto3.getEnumType(RacePhase) },
    { no: 8, name: "track", kind: "enum", T: proto3.getEnumType(Track) },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): SaveFrameClassificationRequest {
    return new SaveFrameClassificationRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): SaveFrameClassificationRequest {
    return new SaveFrameClassificationRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): SaveFrameClassificationRequest {
    return new SaveFrameClassificationRequest().fromJsonString(jsonString, options);
  }

  static equals(a: SaveFrameClassificationRequest | PlainMessage<SaveFrameClassificationRequest> | undefined, b: SaveFrameClassificationRequest | PlainMessage<SaveFrameClassificationRequest> | undefined): boolean {
    return proto3.util.equals(SaveFrameClassificationRequest, a, b);
  }
}

/**
 * 2. Getting an okay back
 *
 * @generated from message mkw.shared.SaveFrameClassificationResponse
 */
export class SaveFrameClassificationResponse extends Message<SaveFrameClassificationResponse> {
  /**
   * @generated from field: bool success = 1;
   */
  success = false;

  constructor(data?: PartialMessage<SaveFrameClassificationResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "mkw.shared.SaveFrameClassificationResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "success", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): SaveFrameClassificationResponse {
    return new SaveFrameClassificationResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): SaveFrameClassificationResponse {
    return new SaveFrameClassificationResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): SaveFrameClassificationResponse {
    return new SaveFrameClassificationResponse().fromJsonString(jsonString, options);
  }

  static equals(a: SaveFrameClassificationResponse | PlainMessage<SaveFrameClassificationResponse> | undefined, b: SaveFrameClassificationResponse | PlainMessage<SaveFrameClassificationResponse> | undefined): boolean {
    return proto3.util.equals(SaveFrameClassificationResponse, a, b);
  }
}